Practical 1

+++ function-1-2.cpp
@@ -7,7 +7,7 @@ double array_mean(int array[], int n) {
     for (int i = 0; i < n; i++) {
         sum = sum + array[i];
     }
-    mean = sum / n;
+    mean = double(sum /n);
     return mean;
 
 }
